fix: pt atomic models

+++ deepmd/pt/model/atomic_model/pairtab_atomic_model.py
@@ -269,9 +269,11 @@ class PairTabAtomicModel(BaseAtomicModel):
         # i_type : (nframes, nloc), this is atype.
         # j_type : (nframes, nloc, nnei)
         j_type = extended_atype[
-            torch.arange(extended_atype.size(0), device=extended_coord.device)[  # pylint: disable=no-explicit-dtype
-                :, None, None
-            ],
+            torch.arange(
+                extended_atype.size(0),
+                device=extended_coord.device,
+                dtype=extended_atype.dtype,
+            )[:, None, None],
             masked_nlist,
         ]
